More stable parsing of `mtr` JSON output
https://github.com/traviscross/mtr/pull/355 changed the type of
some fields from "number as string" to a JSON number. This became
effective with mtr:0.94.

Some environments (such as Ubuntu:20.04) still have mtr:0.93 and below
available. Instead of discussing the "urge to upgrade" and spend breath
on that, this commit uses the power of json.UnmarshalJSON() from Golang
to be a little bit more flexible on the topic. So, mtr-exporter now
supports both mtr:0.93 numbers like `"0x0"` and the mtr:0.94+ equivalent
`0`.

+++ pkg/job/prometheus.go
@@ -75,7 +75,7 @@ func (c *Collector) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 
 		for i, hub := range report.Hubs {
 			labels["host"] = hub.Host
-			labels["count"] = strconv.FormatInt(hub.Count, integerBase)
+			labels["count"] = strconv.FormatInt(int64(hub.Count), integerBase)
 			// mark last hub to have it easily identified
 			if i < (len(report.Hubs) - 1) {
 				hub.WriteMetrics(w, labels2Prom(labels), tsMs)
